+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/views/MapView.java
@@ -2314,8 +2314,8 @@ public final class MapView extends FrameLayout {
 
         mNativeMapView.addAnnotationIcon(
                 id,
-                (int) (bitmap.getWidth() / scale),
-                (int) (bitmap.getHeight() / scale),
+                bitmap.getWidth(),
+                bitmap.getHeight(),
                 scale, buffer.array());
     }